Fix : authentication name 불일치 수정
- 사용자 인증에 필요한 Autentication name이 일치하지 않던 오류를 수정했습니다.
- 기존 사용자가 로그인 할 경우 기존 kakaoId로 인증 객체를 생성하던것을 userUUID로 변경하였습니다.
- 이전 커밋에서 필요없는 코드를 제거하였습니다.

- 토큰의 subject와 Authentication의 name을 일치시켰습니다.

+++ src/main/java/shootingstar/var/jwt/TokenProperty.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Component;
 @Getter
 public class TokenProperty {
 //    private final int ACCESS_EXPIRE = 3 * 60 * 60 * 1000 ; // 3시간 밀리 세컨드
-    private final int ACCESS_EXPIRE = 30 * 24 * 60 * 60 * 1000 ; // 30일 밀리 세컨드
+    private final long ACCESS_EXPIRE = 30L * 24 * 60 * 60 * 1000 ; // 30일 밀리 세컨드
 //    private final int REFRESH_EXPIRE = 24 * 60 * 60 * 1000 ; // 24시간 밀리 세컨드
     private final int REFRESH_EXPIRE = 24 * 60 * 60 * 1000 ; // 24시간 밀리 세컨드
 }
